refactor: shrink dependency graph + hide internal show flags
Three small architectural cuts identified in the 0.6.0 review (#177).

- **`cosineSimilarity` moved to `src/embedders/types.ts`** alongside
  `EmbeddingVector`. `db.ts` now imports both from the lightweight types
  module instead of pulling in the embedder facade and its
  `@huggingface/transformers` chain just for one math helper.
  `src/embedder.ts` re-exports for backwards compatibility.
- **Indexer no longer re-exports `buildSearchText` / `buildSearchFields`.**
  `db-search.ts` now imports directly from `./search-fields` so any code
  path that needs text-building does not transitively load the indexer's
  LLM and embedder dependencies. Test files updated to match.
- **Internal `--akmView` / `--akmHeading` / `--akmStart` / `--akmEnd`
  flags hidden from `akm show --help`.** These are populated by
  `normalizeShowArgv` from positional knowledge-view syntax and were
  never meant for direct user use. Read via `parseFlagValue` from
  `process.argv` instead of citty args, so they never surface in help.

`akm show --help` now shows only `--format` and `--detail`.

bun test → 1650 pass / 0 fail / 7 skip. tsc + biome clean.
