+++ src/embedder.ts
@@ -100,26 +100,11 @@ export async function embedBatch(
 
 // ── Similarity ──────────────────────────────────────────────────────────────
 
-export function cosineSimilarity(a: EmbeddingVector, b: EmbeddingVector): number {
-  if (a.length !== b.length) {
-    // MD-4: Return 0 on dimension mismatch rather than silently computing on a
-    // truncated view, which would produce meaningless similarity scores.
-    warn("cosineSimilarity: vector dimension mismatch (%d vs %d) — re-index recommended", a.length, b.length);
-    return 0;
-  }
-  const len = a.length;
-  if (len === 0) return 0;
-  let dot = 0,
-    magA = 0,
-    magB = 0;
-  for (let i = 0; i < len; i++) {
-    dot += a[i] * b[i];
-    magA += a[i] * a[i];
-    magB += b[i] * b[i];
-  }
-  const denom = Math.sqrt(magA) * Math.sqrt(magB);
-  return denom === 0 ? 0 : dot / denom;
-}
+// `cosineSimilarity` was moved to `./embedders/types.ts` so importers
+// (notably `db.ts`) can pull the math function without dragging in this
+// facade and its `@huggingface/transformers` import chain. Re-export
+// preserves the existing public API.
+export { cosineSimilarity } from "./embedders/types";
 
 // ── Availability check ──────────────────────────────────────────────────────
