+++ tests/benchmark-suite.ts
@@ -20,9 +20,10 @@ import os from "node:os";
 import path from "node:path";
 import { saveConfig } from "../src/config";
 import { closeDatabase, openDatabase, rebuildFts, upsertUtilityScore } from "../src/db";
-import { buildSearchFields, recomputeUtilityScores } from "../src/indexer";
+import { recomputeUtilityScores } from "../src/indexer";
 import { assembleInfo } from "../src/info";
 import { getDbPath } from "../src/paths";
+import { buildSearchFields } from "../src/search-fields";
 import { akmSearch } from "../src/stash-search";
 import type { StashSearchHit } from "../src/stash-types";
 import { insertUsageEvent } from "../src/usage-events";
